refactor: moved accesstoken, passwordresettoken test to authentication module
refactor: exposed listroutes index in documentation
