+++ src/apps/backend/modules/documentation/index.ts
@@ -1,3 +1,5 @@
 export { default as DocumentationService } from 'backend/modules/documentation/documentation-service';
 export { default as DocumentationServer } from 'backend/modules/documentation/rest-api/documentation-server';
 export * from 'backend/modules/documentation/types';
+export { default as expressListRoutes } from 'backend/modules/documentation/list-routes/express-list-routes';
+export * from 'backend/modules/documentation/list-routes/types';
