+++ src/apps/backend/modules/documentation/internals/document-generator-util.ts
@@ -5,8 +5,8 @@ import {
   ErrorReadingFile,
   GetRoutesWithControllerAndSerializerDetailsResponse,
   HttpRouteWithControllerAndSerializerDetails,
+  HttpRoute,
 } from 'backend/modules/documentation';
-import { HttpRoute } from 'backend/modules/documentation/list-routes';
 
 export default class DocumentGeneratorUtil {
   public static controllerFileSuffix = '-controller.js';
