--- src/apps/backend/modules/documentation/list-routes/index.ts
@@ -1,2 +1,0 @@
-export { default as expressListRoutes } from 'backend/modules/documentation/list-routes/express-list-routes';
-export * from 'backend/modules/documentation/list-routes/types';
